// SIP/SIP.Experts/src/enums/EITAppWindow.ts
enum EITAppWindow {
    map = "map",
    tools = "tools",
    tables = "tables",
    coordList = "coordList",
    cart = "cart",
    testTools = "testTools",
    layers = "layers",
    legend = "legend",
    print = "print",
    feedback = "feedback",
    detailEditing = "detailEditing",
    organizationEditing = "organizationEditing",
    servicePreview = "servicePreview",
    adService1Wizard = "adService1Wizard",
    adService2Wizard = "adService2Wizard",
    adService3Wizard = "adService3Wizard",
    adService4Wizard = "adService4Wizard",
    adService51Wizard = "adService51Wizard",
    adService52Wizard = "adService52Wizard",
    adService53Wizard = "adService53Wizard",
    adService1Info = "adService1Info",
    adService2Info = "adService2Info",
    adService3Info = "adService3Info",
    adService4Info = "adService4Info",
    adService51Info = "adService51Info",
    adService52Info = "adService52Info",
    adService53Info = "adService53Info",
    importData = "importData",
    news = "news",
    contentDoc = "contentDoc",
    openData = "openData",
    inspire = "inspire",
    metaDatSearch = "metaDatSearch",
    createUserRequest = "createUserRequest",
    bulletin = "bulletin"
}

export default EITAppWindow;